+++ libs/react/src/components/molecules/percentSlider/percentSlider.stories.tsx
@@ -14,7 +14,7 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  render: (args) => {
+  render: args => {
     // eslint-disable-next-line react-hooks/rules-of-hooks
     const [value, setValue] = React.useState<number | null>(args.value);