+++ libs/react/src/components/organisms/complianceWizardLair/surveyComplianceFlowStep/surveyComplianceFlowStep.tsx
@@ -1,13 +1,26 @@
 import React, { useContext, useEffect } from 'react';
-import { ComplianceSurveyLeftNav, ComplianceSurveyQuestionnaire, ComplianceSurveyRightNav, ErrorFallback, Spinner, Takeover, WizardContext } from '@coldpbc/components';
-import { ComplianceSurveyActiveKeyType, ComplianceSurveyPayloadType } from '@coldpbc/interfaces';
-import { getStartingKey, sortComplianceSurvey } from '@coldpbc/lib';
+import {
+  ComplianceSurveyLeftNav,
+  ComplianceSurveyQuestionnaire,
+  ComplianceSurveyRightNav,
+  ComplianceSurveySavedQuestionnaire,
+  ErrorFallback,
+  Spinner,
+  Takeover,
+  WizardContext,
+} from '@coldpbc/components';
+import { ComplianceSurveyActiveKeyType, ComplianceSurveyPayloadType, ComplianceSurveySavedQuestionType } from '@coldpbc/interfaces';
+import { getAccurateBookmarkedValue, getQuestionValue, getSavedQuestionsInSurvey, getStartingKey, putSurveyData, sortComplianceSurvey, updateSurveyQuestion } from '@coldpbc/lib';
 import { GlobalSizes } from '@coldpbc/enums';
 import { withErrorBoundary } from 'react-error-boundary';
+import { useAuth0Wrapper } from '@coldpbc/hooks';
+import { get, set } from 'lodash';
+import { mutate } from 'swr';
 
 const _SurveyComplianceFlowStep = () => {
   const { data } = useContext(WizardContext);
-  const { surveyData } = data as { surveyData: ComplianceSurveyPayloadType };
+  const { orgId, getOrgSpecificUrl } = useAuth0Wrapper();
+  const { surveyData, name } = data as { surveyData: ComplianceSurveyPayloadType; name: string };
   const [activeKey, setActiveKey] = React.useState<ComplianceSurveyActiveKeyType>({
     value: '',
     section: '',
@@ -17,24 +30,60 @@ const _SurveyComplianceFlowStep = () => {
   });
   const [surveyOpen, setSurveyOpen] = React.useState(false);
   const sortedSurvey = sortComplianceSurvey(surveyData);
+  const originalSavedQuestionsState = getSavedQuestionsInSurvey(sortedSurvey);
   const [surveyState, setSurveyState] = React.useState<ComplianceSurveyPayloadType>(sortedSurvey);
+  const [savedQuestions, setSavedQuestions] = React.useState<Array<ComplianceSurveySavedQuestionType>>(originalSavedQuestionsState);
+  const [bookmarked, setBookmarked] = React.useState<{
+    [key: string]: boolean;
+  }>({});
+
+  const onSurveyClose = async () => {
+    if (activeKey.section === 'savedQuestions') {
+      const bookmarkedQuestion = getAccurateBookmarkedValue(surveyState.definition.sections, activeKey, bookmarked);
+      const newSurvey = updateSurveyQuestion(surveyData, activeKey, {
+        value: getQuestionValue(surveyData, activeKey),
+        skipped: false,
+        saved: bookmarkedQuestion,
+      });
+      const response = (await putSurveyData(newSurvey as ComplianceSurveyPayloadType, getOrgSpecificUrl)) as ComplianceSurveyPayloadType;
+      const sortedSurvey = sortComplianceSurvey(response);
+      setSurveyState(sortedSurvey);
+      await mutate([getOrgSpecificUrl(`/surveys/${newSurvey.name}`), 'GET'], sortedSurvey, {
+        revalidate: false,
+      });
+    }
+  };
 
   useEffect(() => {
-    // set start active key
-    if (!surveyData) {
+    if (!surveyData || !orgId) {
       return;
     }
+    // runs when the component mounts and/or when the orgId changes via impersonation
     const key = getStartingKey(sortedSurvey);
-    const activeKeyFromStorage = localStorage.getItem('complianceActiveKey');
-    const keyToBeUsed = activeKeyFromStorage ? JSON.parse(activeKeyFromStorage) : key;
+    const orgStorage = localStorage.getItem(orgId);
+    const parsedOrgStorage = orgStorage ? JSON.parse(orgStorage) : {};
+    const activeKeyFromStorage = get(parsedOrgStorage, `compliance.${name}.complianceActiveKey`, null);
+    const keyToBeUsed = activeKeyFromStorage ? activeKeyFromStorage : key;
     setActiveKey(keyToBeUsed);
-  }, []);
+  }, [orgId]);
 
   useEffect(() => {
-    if (activeKey.section !== '') {
-      localStorage.setItem('complianceActiveKey', JSON.stringify(activeKey));
+    if (activeKey.section !== '' && orgId) {
+      const orgStorage = localStorage.getItem(orgId);
+      const parsedOrgStorage = orgStorage ? JSON.parse(orgStorage) : {};
+      set(parsedOrgStorage, `compliance.${name}.complianceActiveKey`, activeKey);
+      localStorage.setItem(orgId, JSON.stringify(parsedOrgStorage));
     }
   }, [activeKey]);
+
+  useEffect(() => {
+    const newSavedQuestions = getSavedQuestionsInSurvey(surveyState);
+    setSavedQuestions(newSavedQuestions);
+    if (newSavedQuestions.length === 0 && activeKey.section === 'savedQuestions') {
+      const key = getStartingKey(sortedSurvey);
+      setActiveKey(key);
+    }
+  }, [surveyState]);
 
   if (!sortedSurvey) {
     return null;
@@ -47,33 +96,45 @@ const _SurveyComplianceFlowStep = () => {
   return (
     <div className={'h-[708px] w-full flex flex-row relative text-tc-primary'}>
       <div className={'flex w-auto h-full'}>
-        <ComplianceSurveyLeftNav complianceSet={surveyState} activeKey={activeKey} setActiveKey={setActiveKey} />
+        <ComplianceSurveyLeftNav surveyData={surveyState} savedQuestions={savedQuestions} activeKey={activeKey} setActiveKey={setActiveKey} />
       </div>
       <div className={'flex w-full h-full'}>
-        <ComplianceSurveyRightNav surveyData={surveyState} activeKey={activeKey} setActiveKey={setActiveKey} surveyOpen={surveyOpen} setSurveyOpen={setSurveyOpen} />
-      </div>
-      <Takeover
-        show={surveyOpen}
-        setShow={setSurveyOpen}
-        className={'absolute h-full w-full p-[30px] bg-transparent'}
-        containClassName={'bg-bgc-elevated rounded-lg'}
-        header={{
-          dismiss: {
-            dismissible: true,
-            onClick: () => {
-              setSurveyOpen(false);
-            },
-          },
-        }}>
-        <ComplianceSurveyQuestionnaire
+        <ComplianceSurveyRightNav
           surveyData={surveyState}
-          setSurveyData={setSurveyState}
+          savedQuestions={savedQuestions}
           activeKey={activeKey}
           setActiveKey={setActiveKey}
-          submitSurvey={() => {
-            setSurveyOpen(false);
-          }}
+          surveyOpen={surveyOpen}
+          setSurveyOpen={setSurveyOpen}
         />
+      </div>
+      <Takeover show={surveyOpen} setShow={setSurveyOpen} className={'absolute h-full w-full p-[30px] bg-transparent'} containClassName={'bg-bgc-elevated rounded-lg'}>
+        {activeKey.section === 'savedQuestions' ? (
+          <ComplianceSurveySavedQuestionnaire
+            surveyData={surveyState}
+            setSurveyData={setSurveyState}
+            activeKey={activeKey}
+            setActiveKey={setActiveKey}
+            submitSurvey={() => {
+              onSurveyClose();
+              setSurveyOpen(false);
+            }}
+            savedQuestions={savedQuestions}
+            bookmarked={bookmarked}
+            setBookmarked={setBookmarked}
+          />
+        ) : (
+          <ComplianceSurveyQuestionnaire
+            surveyData={surveyState}
+            setSurveyData={setSurveyState}
+            activeKey={activeKey}
+            setActiveKey={setActiveKey}
+            submitSurvey={() => {
+              setSurveyOpen(false);
+            }}
+            savedQuestions={savedQuestions}
+          />
+        )}
       </Takeover>
     </div>
   );